+++ CHANGELOG.md
@@ -2,6 +2,8 @@ UNRELEASED
 ----------
 
 * [ [#1454](https://github.com/digitalfabrik/integreat-cms/issues/1454) ] Enable recurring events for non-expert users
+* [ [#1416](https://github.com/digitalfabrik/integreat-cms/issues/1416) ] Hide staff users from region user list
+* [ [#1483](https://github.com/digitalfabrik/integreat-cms/issues/1483) ] Add filters to admin user list
 
 
 2022.5.3